Add files via upload
This is a basic coffee machine created through python. 
Gives the option of ordering a coffee, namely: Espresso, Latte, and Cappuccino. 
A report of how much water, milk, and coffee are left is also displayed just by typing the keyword, "report".
Money paid by the customer can also be viewed in this report.